requirement.txt
adbutils==0.7.1
allure-pytest==2.8.12
allure-python-commons==2.8.12
altgraph==0.16.1
apkutils2==1.0.0
Appium-Python-Client==0.50
APScheduler==3.6.1
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.0
bottle==0.12.17
bs4==0.0.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cigam==0.0.3
Click==7.0
colorama==0.4.1
cryptography==2.7
decorator==4.4.2
demjson==2.2.4
Deprecated==1.2.7
deprecation==2.0.7
Django==2.2.5
django-bootstrap3==11.1.0
Flask==1.1.1
Flask-WTF==0.14.3
future==0.18.2
gevent==1.4.0
greenlet==0.4.15
humanize==1.0.0
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.1.1
logzero==1.5.0
lxml==4.4.2
MarkupSafe==1.1.1
more-itertools==7.2.0
nose==1.3.7
numpy==1.17.3
packaging==19.2
pandas==0.25.2
paramiko==2.6.0
pefile==2019.4.18
Pillow==7.0.0
pluggy==0.13.0
prettytable==0.7.2
progress==1.5
public==2019.4.13
py==1.8.0
pycparser==2.19
pyecharts==1.5.1
pyelftools==0.26
PyExecJS==1.5.1
PyInstaller==3.5
PyMySQL==0.9.3
PyNaCl==1.3.0
pyparsing==2.4.2
pyrsistent==0.15.4
pytest==5.4.1
python-dateutil==2.8.0
pytz==2019.2
pywin32-ctypes==0.2.0
PyYAML==5.3
requests==2.22.0
retry==0.9.2
selenium==3.141.0
self==2019.4.13
simplejson==3.16.0
six==1.12.0
soupsieve==1.9.3
sqlparse==0.3.0
tzlocal==2.0.0
uiautomator2==2.5.8.dev1
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.16.0
whichcraft==0.6.1
wrapt==1.12.0
WTForms==2.2.1
xlrd==1.2.0
XlsxWriter==1.2.2
xlutils==2.0.0
xlwt==1.3.0
xmltodict==0.12.0
zipp==0.6.0
